requirements.txt
# Discord Bot Requirements
# Python dependencies for Discord Member Monitoring Bot

# Core Discord library
discord.py>=2.3.0

# Environment and configuration
python-dotenv>=1.0.0
PyYAML>=6.0

# Async file operations
aiofiles>=23.0.0

# Async SQLite database
aiosqlite>=0.19.0

# HTTP client for Discord API
aiohttp>=3.8.0

# Console output formatting
colorama>=0.4.6

# Additional utilities
typing-extensions>=4.0.0
